refactor: move methods from `ProjectService` to `ProjectPaperService`
