+++ src/main/kotlin/se/uulm/snowballr/backend/service/ProjectService.kt
@@ -3,46 +3,24 @@ package se.uulm.snowballr.backend.service
 import se.uulm.snowballr.backend.grpc.SnowballRServer.SnowballRService
 import se.uulm.snowballr.backend.model.AccessType
 import se.uulm.snowballr.backend.model.EntityType
-import se.uulm.snowballr.backend.model.SnowballRException.DuplicateEntityException
 import se.uulm.snowballr.backend.model.SnowballRException.FailedPreconditionException
-import se.uulm.snowballr.backend.model.SnowballRException.NotFoundException
-import se.uulm.snowballr.backend.model.SnowballRException.OutOfRangeException
 import se.uulm.snowballr.backend.model.SnowballRException.UnauthorizedException
-import se.uulm.snowballr.backend.model.dto.Paper
-import se.uulm.snowballr.backend.model.dto.Project
-import se.uulm.snowballr.backend.model.dto.ProjectPaper
-import se.uulm.snowballr.backend.model.dto.ProjectPaperWithPaper
 import se.uulm.snowballr.backend.model.dto.User
-import se.uulm.snowballr.backend.model.dto.toGrpcAuthor
 import se.uulm.snowballr.backend.model.dto.toGrpcProject
 import se.uulm.snowballr.backend.model.dto.toGrpcProjectMembers
-import se.uulm.snowballr.backend.model.dto.toGrpcProjectPaper
-import se.uulm.snowballr.backend.model.dto.toGrpcProjectPapers
 import se.uulm.snowballr.backend.model.dto.toGrpcProjects
-import se.uulm.snowballr.backend.model.dto.toGrpcReview
 import se.uulm.snowballr.backend.model.parseUUID
 import se.uulm.snowballr.backend.repository.ICriterionTableRepo
-import se.uulm.snowballr.backend.repository.IPaperTableRepo
 import se.uulm.snowballr.backend.repository.IProjectTableRepo
-import se.uulm.snowballr.backend.repository.IReviewTableRepo
 import se.uulm.snowballr.backend.repository.IUserTableRepo
-import se.uulm.snowballr.backend.repository.association.IAuthorOfPaperTableRepo
-import se.uulm.snowballr.backend.repository.association.ICitationTableRepo
 import se.uulm.snowballr.backend.repository.association.IProjectMemberTableRepo
-import se.uulm.snowballr.backend.repository.association.IProjectPaperTableRepo
-import se.uulm.snowballr.backend.repository.association.IReviewHasCriterionTableRepo
 import snowballr.Base
-import snowballr.ProjectOuterClass.PaperDecision
 import snowballr.ProjectOuterClass.ProjectStatus
 import java.util.UUID
 import snowballr.CriterionOuterClass.Criterion as GrpcCriterion
-import snowballr.PaperOuterClass.Author as GrpcAuthor
 import snowballr.ProjectOuterClass.Project as GrpcProject
 import snowballr.ProjectOuterClass.Project.Member as GrpcProjectMember
-import snowballr.ProjectOuterClass.Project.Paper as GrpcProjectPaper
-import snowballr.ReviewOuterClass.Review as GrpcReview
 
-@Suppress("ComplexInterface", "TooManyFunctions")
 interface IProjectService {
     /**
      * Service implementation of [SnowballRService.getProjectById].
@@ -82,32 +60,7 @@ interface IProjectService {
     /**
      * Service implementation of [SnowballRService.getProjectMembers].
      */
-    suspend fun getProjectMembers(request: Base.Id): GrpcProjectMember.List
-
-    /**
-     * Service implementation of [SnowballRService.getProjectPaperById].
-     */
-    suspend fun getProjectPaperById(request: Base.Id): GrpcProjectPaper
-
-    /**
-     * Service implementation of [SnowballRService.getProjectPaperByRelativeId].
-     */
-    suspend fun getProjectPaperByRelativeId(request: GrpcProjectPaper.Get): GrpcProjectPaper
-
-    /**
-     * Service implementation of [SnowballRService.getAllProjectPapersForProject].
-     */
-    suspend fun getAllProjectPapersForProject(request: Base.Id): GrpcProjectPaper.List
-
-    /**
-     * Service implementation of [SnowballRService.getPapersToReviewForProject].
-     */
-    suspend fun getPapersToReviewForProject(request: Base.Id): GrpcProjectPaper.List
-
-    /**
-     * Service implementation of [SnowballRService.addPaperToProject].
-     */
-    suspend fun addPaperToProject(request: GrpcProjectPaper.Add): GrpcProjectPaper
+    suspend fun getProjectMembers(request: Base.Id): GrpcProject.Member.List
 }
 
 /**
@@ -121,40 +74,13 @@ interface IProjectService {
  * @param userRepo The repository responsible for managing persistence operations for users.
  * @param projectMemberRepo The repository responsible for managing persistence operations for project members.
  * @param criterionRepo The repository responsible for managing persistence operations for criteria.
- * @param paperRepo The repository responsible for managing persistence operations for papers.
- * @param projectPaperRepo The repository responsible for managing persistence operations for project papers.
- * @param authorOfPaperTableRepo The repository responsible for managing persistence operations for the author
- * paper relation.
- * @param citationTableRepo The repository responsible for managing persistence operations for the citation relation.
- * @param reviewTableRepo The repository responsible for managing persistence operations for the reviews
- * @param reviewHasCriterionTableRepo The repository responsible for managing persistence operations for the review has
- * criterion relation.
  */
-@Suppress("LongParameterList", "TooManyFunctions")
 class ProjectService(
     private val repo: IProjectTableRepo,
     private val userRepo: IUserTableRepo,
     private val projectMemberRepo: IProjectMemberTableRepo,
     private val criterionRepo: ICriterionTableRepo,
-    private val paperRepo: IPaperTableRepo,
-    private val projectPaperRepo: IProjectPaperTableRepo,
-    private val authorOfPaperTableRepo: IAuthorOfPaperTableRepo,
-    private val citationTableRepo: ICitationTableRepo,
-    private val reviewTableRepo: IReviewTableRepo,
-    private val reviewHasCriterionTableRepo: IReviewHasCriterionTableRepo,
 ) : IProjectService {
-    /**
-     * Checks whether a user is a member of a specific project.
-     *
-     * @param projectId The unique identifier of the project.
-     * @param currentUserId The unique identifier of the user to be checked for membership.
-     * @return `true` if the user is a member of the project, `false` otherwise.
-     */
-    private suspend fun isProjectMember(projectId: UUID, currentUserId: UUID): Boolean {
-        val projectMembers = projectMemberRepo.getProjectMembers(projectId)
-        return projectMembers.any { it.userId == currentUserId }
-    }
-
     private suspend fun ensureCurrentUserIsProjectMember(projectId: UUID, currentUser: User) {
         if (isProjectMember(projectId, currentUser.id)) return
 
@@ -162,108 +88,6 @@ class ProjectService(
             throw UnauthorizedException.Single(EntityType.PROJECT, projectId.toString(), AccessType.READ, it)
         }
     }
-
-    /**
-     * Converts a combination of [Paper] and [ProjectPaper] into a gRPC-compatible [GrpcProjectPaper] object.
-     * This includes enriching the data with associated authors, backward references, and reviews.
-     *
-     * @param paper The [Paper] containing base paper details.
-     * @param projectPaper The [ProjectPaper] linking the [Paper] to a specific project and its metadata.
-     * @return A [GrpcProjectPaper] object that encapsulates the paper, associated metadata, authors,
-     * backward references, and reviews.
-     */
-    private suspend fun getGrpcProjectPaper(paper: Paper, projectPaper: ProjectPaper): GrpcProjectPaper {
-        val authors = authorOfPaperTableRepo.getAuthorsOfPaperById(paper.id).map { it.toGrpcAuthor() }
-        val backwardReferences = citationTableRepo
-            .getBackwardsReferencedPaperIdsOfPaperById(paper.id).map(UUID::toString)
-        val reviews = reviewTableRepo.getAllReviewsForProjectPaper(projectPaper.id)
-            .map {
-                val selectedCriteriaIds = reviewHasCriterionTableRepo.getSelectedCriteriaIdsForReviewById(it.id)
-                it.toGrpcReview(selectedCriteriaIds.map(UUID::toString))
-            }
-        return ProjectPaperWithPaper(projectPaper, paper).toGrpcProjectPaper(authors, backwardReferences, reviews)
-    }
-
-    /**
-     * Retrieves a list of [ProjectPaper] associated with a specified project id. This method also ensures access
-     * control for the current user. Optionally, a predicate function can be provided to filter the [ProjectPaper]s
-     * based on custom criteria.
-     *
-     * @param request The request containing the ID of the [Project] for which [ProjectPaper]s are to be retrieved.
-     * @param predicate An optional lambda function that takes a [ProjectPaperWithPaper], a map of [GrpcReview], and a
-     * user ID. This function should return a boolean value to filter the [ProjectPaperWithPaper]s. If null, no
-     * filtering is applied.
-     * @return A list of [GrpcProjectPaper] including associated metadata such as authors, backward references, and
-     * reviews.
-     * @throws UnauthorizedException If the user does not have the required access to the project.
-     */
-    private suspend fun getProjectPapers(
-        request: Base.Id,
-        predicate: (suspend (ProjectPaperWithPaper, Map<ProjectPaper, List<GrpcReview>>, String) -> Boolean)? = null,
-    ): GrpcProjectPaper.List = withUser(userRepo) { currentUser ->
-        val projectId = parseUUID(request.id, EntityType.PROJECT)
-        if (!repo.doesProjectExistById(projectId)) {
-            throw NotFoundException(EntityType.PROJECT, projectId.toString())
-        }
-
-        ensureCurrentUserIsProjectMember(projectId, currentUser)
-
-        var projectPapersWithPapers = projectPaperRepo.getAllProjectPapersWithPapers(projectId)
-        val paperAuthorsMap = mutableMapOf<Paper, List<GrpcAuthor>>()
-        val paperBackwardReferencesMap = mutableMapOf<Paper, List<String>>()
-        val projectPaperReviewsMap = mutableMapOf<ProjectPaper, List<GrpcReview>>()
-        for (projectPaper in projectPapersWithPapers) {
-            val paper = projectPaper.paper
-            paperAuthorsMap[paper] = authorOfPaperTableRepo
-                .getAuthorsOfPaperById(paper.id).map { it.toGrpcAuthor() }
-            paperBackwardReferencesMap[paper] = citationTableRepo
-                .getBackwardsReferencedPaperIdsOfPaperById(paper.id).map(UUID::toString)
-            projectPaperReviewsMap[projectPaper.projectPaper] = reviewTableRepo
-                .getAllReviewsForProjectPaper(projectPaper.projectPaper.id)
-                .map {
-                    val selectedCriteriaIds = reviewHasCriterionTableRepo.getSelectedCriteriaIdsForReviewById(it.id)
-                    it.toGrpcReview(selectedCriteriaIds.map(UUID::toString))
-                }
-        }
-        projectPapersWithPapers = predicate?.let { pred ->
-            projectPapersWithPapers.filter { pred(it, projectPaperReviewsMap, currentUser.id.toString()) }
-        } ?: projectPapersWithPapers
-
-        projectPapersWithPapers.toGrpcProjectPapers(paperAuthorsMap, paperBackwardReferencesMap, projectPaperReviewsMap)
-    }
-
-    /**
-     * Retrieves the gRPC representation of a project's paper, including associated data such as authors,
-     * backward references, and reviews. Furthermore, it ensures that the current user has the necessary permissions to access
-     * the project's paper.
-     *
-     * @param projectPaper The project paper, containing the ids linking to the associated data.
-     * @param projectId The unique identifier of the project to which the paper belongs.
-     * @return The gRPC representation of the project's paper, including associated data.
-     * @throws UnauthorizedException.Single If the current user does not have the required read access to the project.
-     */
-    private suspend fun loadAndAuthorizeProjectPaper(projectPaper: ProjectPaper, projectId: UUID): GrpcProjectPaper =
-        withUser(userRepo) { currentUser ->
-            ensureCurrentUserIsProjectMember(projectId, currentUser)
-
-            val paper = paperRepo.getPaperById(projectPaper.paperId).getOrThrow()
-            val authors = authorOfPaperTableRepo
-                .getAuthorsOfPaperById(paper.id)
-                .map { it.toGrpcAuthor() }
-
-            val backwardReferences = citationTableRepo
-                .getBackwardsReferencedPaperIdsOfPaperById(paper.id)
-                .map { it.toString() }
-
-            val reviews = reviewTableRepo
-                .getAllReviewsForProjectPaper(projectPaper.id)
-                .map {
-                    val selectedCriteriaIds = reviewHasCriterionTableRepo.getSelectedCriteriaIdsForReviewById(it.id)
-                    it.toGrpcReview(selectedCriteriaIds.map(UUID::toString))
-                }
-
-            ProjectPaperWithPaper(projectPaper, paper).toGrpcProjectPaper(authors, backwardReferences, reviews)
-        }
 
     override suspend fun getProjectById(request: Base.Id): GrpcProject = withUser(userRepo) { currentUser ->
         val projectId = parseUUID(request.id, EntityType.PROJECT)
@@ -356,61 +180,5 @@ class ProjectService(
             }
 
             projectMembersWithUsers.toGrpcProjectMembers()
-        }
-
-    override suspend fun getProjectPaperById(request: Base.Id): GrpcProjectPaper {
-        val projectPaperId = parseUUID(request.id, EntityType.PROJECT_PAPER)
-        val projectPaper = projectPaperRepo.getProjectPaperById(projectPaperId).getOrThrow()
-        return loadAndAuthorizeProjectPaper(projectPaper, projectPaper.projectId)
-    }
-
-    override suspend fun getProjectPaperByRelativeId(request: GrpcProjectPaper.Get): GrpcProjectPaper {
-        val projectId = parseUUID(request.projectId, EntityType.PROJECT)
-        if (!repo.doesProjectExistById(projectId)) {
-            throw NotFoundException(EntityType.PROJECT, projectId.toString())
-        }
-
-        val relativeId = request.relativeProjectPaperId.toLong()
-        val projectPaper = projectPaperRepo.getProjectPaperByRelativeId(projectId, relativeId).getOrThrow()
-        return loadAndAuthorizeProjectPaper(projectPaper, projectId)
-    }
-
-    override suspend fun getAllProjectPapersForProject(request: Base.Id): GrpcProjectPaper.List =
-        getProjectPapers(request)
-
-    override suspend fun getPapersToReviewForProject(request: Base.Id): GrpcProjectPaper.List {
-        val predicate: (ProjectPaperWithPaper, Map<ProjectPaper, List<GrpcReview>>, String) -> Boolean =
-            { projectPaper, projectPaperReviewsMap, currentUserId ->
-                val isAlreadyReviewedByCurrentUser = projectPaperReviewsMap[projectPaper.projectPaper]
-                    ?.any { review -> review.userId == currentUserId } == true
-                val isStillUndecided =
-                    projectPaper.projectPaper.decision == PaperDecision.PAPER_DECISION_UNREVIEWED ||
-                        projectPaper.projectPaper.decision == PaperDecision.PAPER_DECISION_IN_REVIEW
-
-                !isAlreadyReviewedByCurrentUser && isStillUndecided
-            }
-        return getProjectPapers(request, predicate)
-    }
-
-    override suspend fun addPaperToProject(request: GrpcProjectPaper.Add): GrpcProjectPaper =
-        withUser(userRepo) { currentUser ->
-            val projectId = parseUUID(request.projectId, EntityType.PROJECT)
-
-            ensureCurrentUserIsProjectMember(projectId, currentUser)
-
-            val paperId = parseUUID(request.paperId, EntityType.PAPER)
-            val project = repo.getProjectById(projectId).getOrThrow()
-            val paper = paperRepo.getPaperById(paperId).getOrThrow()
-            if (projectPaperRepo.doesProjectPaperExist(projectId, paperId)) {
-                throw DuplicateEntityException(EntityType.PROJECT_PAPER, projectId.toString(), paperId.toString())
-            }
-
-            if (request.stage !in 0..project.maxStage) {
-                throw OutOfRangeException.Stage(request.stage, project.maxStage)
-            }
-
-            val projectPaper = projectPaperRepo.addPaperToProject(request, currentUser.id)
-
-            getGrpcProjectPaper(paper, projectPaper)
         }
 }
